+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* twig-to-php-parser - Update the path for plugin patterns to point to build/patterns.
+
 ## [8.10.0-alpha] - 05-11-2020
 
 ### Added